Reorganise repository, v1.6 first changes
- Reorganise directory structure
- Add optional parameters support
- Fix how some types are displayed when outputting as text
- Use neutral culture when outputting values
