+++ .gitignore
@@ -328,3 +328,15 @@ ASALocalRun/
 
 # MFractors (Xamarin productivity tool) working folder 
 .mfractor/
+
+####################################################################
+
+# Documentation
+*.pdf
+
+# NuGet
+*.nupkg
+
+# Misc ignored files
+_Distr/
+info.txt
