+++ CilBytecodeParser.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Express 2012 for Windows Desktop
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CilBytecodeParserTest", "CilBytecodeParserTest\CilBytecodeParserTest.csproj", "{BB8F6999-8B4C-4339-BA84-6964E5E5EC5A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CilBytecodeParserDemo", "CilBytecodeParserDemo\CilBytecodeParserDemo.csproj", "{BB8F6999-8B4C-4339-BA84-6964E5E5EC5A}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CilBytecodeParser", "CilBytecodeParser\CilBytecodeParser.csproj", "{38496D54-0BC8-479B-A57E-F0EC83E78AD3}"
 EndProject
