+++ CilBytecodeParserDemo/AssemblyInfo.cs
@@ -1,0 +1,22 @@
+﻿/* CilBytecodeParser library 
+ * Copyright (c) 2019,  MSDN.WhiteKnight (https://github.com/MSDN-WhiteKnight) 
+ * License: BSD 2.0 */
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("CilBytecodeParserDemo")]
+[assembly: AssemblyDescription("CIL Bytecode Parser libary demo program")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("CilBytecodeParserTest")]
+[assembly: AssemblyCopyright("Copyright © MSDN.WhiteKnight 2019")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+
+[assembly: ComVisible(false)]
+
+[assembly: Guid("8d3c3ebc-f3f0-4fdc-8c4c-57844d341e5f")]
+
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]
