+++ docfx_project/.gitignore
@@ -7,3 +7,6 @@
 /**/bin/
 /**/obj/
 _site
+
+#############
+*.pdf
