+++ docfx_project/_site_pdf/docfx_project.json
@@ -1,0 +1,1 @@
+{"docfx_project_pdf.pdf":{"docset_name":"docfx_project","asset_id":"pdf/toc.json","toc_files":["pdf/toc.yml"]}}
