+++ docfx_project/api/.manifest
@@ -2,6 +2,7 @@
   "CilBytecodeParser": "CilBytecodeParser.yml",
   "CilBytecodeParser.CilAnalysis": "CilBytecodeParser.CilAnalysis.yml",
   "CilBytecodeParser.CilAnalysis.Error": "CilBytecodeParser.CilAnalysis.yml",
+  "CilBytecodeParser.CilAnalysis.EscapeString(System.String)": "CilBytecodeParser.CilAnalysis.yml",
   "CilBytecodeParser.CilAnalysis.GetGraph(System.Reflection.MethodBase)": "CilBytecodeParser.CilAnalysis.yml",
   "CilBytecodeParser.CilAnalysis.GetReferencedMembers(System.Reflection.Assembly)": "CilBytecodeParser.CilAnalysis.yml",
   "CilBytecodeParser.CilAnalysis.GetReferencedMembers(System.Reflection.Assembly,CilBytecodeParser.MemberCriteria)": "CilBytecodeParser.CilAnalysis.yml",
@@ -27,7 +28,10 @@
   "CilBytecodeParser.CilGraph.#ctor(CilBytecodeParser.CilGraphNode,System.Reflection.MethodBase)": "CilBytecodeParser.CilGraph.yml",
   "CilBytecodeParser.CilGraph._Method": "CilBytecodeParser.CilGraph.yml",
   "CilBytecodeParser.CilGraph._Root": "CilBytecodeParser.CilGraph.yml",
+  "CilBytecodeParser.CilGraph.EmitTo(System.Reflection.Emit.ILGenerator,System.Func{CilBytecodeParser.CilInstruction,System.Boolean})": "CilBytecodeParser.CilGraph.yml",
   "CilBytecodeParser.CilGraph.Error": "CilBytecodeParser.CilGraph.yml",
+  "CilBytecodeParser.CilGraph.GetInstructions": "CilBytecodeParser.CilGraph.yml",
+  "CilBytecodeParser.CilGraph.GetNodes": "CilBytecodeParser.CilGraph.yml",
   "CilBytecodeParser.CilGraph.Method": "CilBytecodeParser.CilGraph.yml",
   "CilBytecodeParser.CilGraph.OnError(System.Object,CilBytecodeParser.CilErrorEventArgs)": "CilBytecodeParser.CilGraph.yml",
   "CilBytecodeParser.CilGraph.Root": "CilBytecodeParser.CilGraph.yml",
@@ -63,6 +67,7 @@
   "CilBytecodeParser.CilInstruction.ByteOffset": "CilBytecodeParser.CilInstruction.yml",
   "CilBytecodeParser.CilInstruction.Code": "CilBytecodeParser.CilInstruction.yml",
   "CilBytecodeParser.CilInstruction.CreateEmptyInstruction(System.Reflection.MethodBase)": "CilBytecodeParser.CilInstruction.yml",
+  "CilBytecodeParser.CilInstruction.EmitTo(System.Reflection.Emit.ILGenerator)": "CilBytecodeParser.CilInstruction.yml",
   "CilBytecodeParser.CilInstruction.Error": "CilBytecodeParser.CilInstruction.yml",
   "CilBytecodeParser.CilInstruction.Method": "CilBytecodeParser.CilInstruction.yml",
   "CilBytecodeParser.CilInstruction.Name": "CilBytecodeParser.CilInstruction.yml",
@@ -72,6 +77,7 @@
   "CilBytecodeParser.CilInstruction.OperandSize": "CilBytecodeParser.CilInstruction.yml",
   "CilBytecodeParser.CilInstruction.OperandType": "CilBytecodeParser.CilInstruction.yml",
   "CilBytecodeParser.CilInstruction.OrdinalNumber": "CilBytecodeParser.CilInstruction.yml",
+  "CilBytecodeParser.CilInstruction.Parse(System.String)": "CilBytecodeParser.CilInstruction.yml",
   "CilBytecodeParser.CilInstruction.ReferencedMember": "CilBytecodeParser.CilInstruction.yml",
   "CilBytecodeParser.CilInstruction.ReferencedString": "CilBytecodeParser.CilInstruction.yml",
   "CilBytecodeParser.CilInstruction.ReferencedType": "CilBytecodeParser.CilInstruction.yml",
@@ -108,6 +114,8 @@
   "CilBytecodeParser.DebugUtils.PrintStackTrace(System.Diagnostics.StackTrace,System.IO.TextWriter)": "CilBytecodeParser.DebugUtils.yml",
   "CilBytecodeParser.Extensions": "CilBytecodeParser.Extensions.yml",
   "CilBytecodeParser.Extensions.CilExtensions": "CilBytecodeParser.Extensions.CilExtensions.yml",
+  "CilBytecodeParser.Extensions.CilExtensions.EmitCilGraph(System.Reflection.Emit.ILGenerator,CilBytecodeParser.CilGraph,System.Func{CilBytecodeParser.CilInstruction,System.Boolean})": "CilBytecodeParser.Extensions.CilExtensions.yml",
+  "CilBytecodeParser.Extensions.CilExtensions.EmitInstruction(System.Reflection.Emit.ILGenerator,CilBytecodeParser.CilInstruction)": "CilBytecodeParser.Extensions.CilExtensions.yml",
   "CilBytecodeParser.Extensions.CilExtensions.GetCilGraph(System.Reflection.MethodBase)": "CilBytecodeParser.Extensions.CilExtensions.yml",
   "CilBytecodeParser.Extensions.CilExtensions.GetCilText(System.Reflection.MethodBase)": "CilBytecodeParser.Extensions.CilExtensions.yml",
   "CilBytecodeParser.Extensions.CilExtensions.GetExecutingInstruction(System.Diagnostics.StackFrame)": "CilBytecodeParser.Extensions.CilExtensions.yml",
