+++ docfx_project/api/CilBytecodeParser.CilErrorEventArgs.yml
@@ -30,6 +30,7 @@ items:
     content: 'public class CilErrorEventArgs : EventArgs'
     content.vb: >-
       Public Class CilErrorEventArgs
+
           Inherits EventArgs
   inheritance:
   - System.Object
