+++ docfx_project/build.cmd
@@ -1,5 +1,5 @@
 rem echo %~dp0
-D:\Distr\Microsoft\docfx\docfx.exe .\docfx.json
+C:\Distr\Microsoft\docfx\docfx.exe .\docfx.json
 copy /Y logo.svg ..\docs
-D:\Distr\Microsoft\docfx\docfx.exe serve ..\docs
+C:\Distr\Microsoft\docfx\docfx.exe serve ..\docs
 pause
